frontend/requirements.txt
Flask==3.0.0
kafka-python==2.0.2
requests==2.31.0